Query the bee taxonomy and country checklist

A simple function to return information about a particular species, including name validity and country occurrences.

Usage

BeeBDCQuery(beeName = NULL, searchChecklist = TRUE, printAllSynonyms = FALSE)

Arguments

beeName

Character or character vector. A single or several bee species names to search for in the beesTaxonomy and beesChecklist tables.

searchChecklist

Logical. If TRUE (default), search the country checklist for each species.

printAllSynonyms

Logical. If TRUE, all synonyms will be printed out for each entered name. default = FALSE.

Value

Returns a list with the elements 'taxonomyReport' and 'SynonymReport'. IF searchChecklist is TRUE, then 'checklistReport' will also be returned.

Examples

  # Single entry example
testQuery <- BeeBDCQuery(
  beeName = "Homalictus fijiensis",
  searchChecklist = TRUE,
  printAllSynonyms = TRUE)
#> Starting taxonomy report...
#> Homalictus fijiensis is a synonym of Lasioglossum fijiense (Perkins and Cheesman, 1928) with the taxon id number 32620.
#>  - 'Homalictus fijiensis' has the synonyms: Halictus fijiensis Perkins and Cheesman, 1928, Halictus suvaensis Cockerell, 1929, Homalictus fijiensis (Perkins and Cheesman, 1928)
#> Starting checklist report...
#>  - Lasioglossum fijiense (Perkins and Cheesman, 1928) is reportedly found in: 
#> Fiji, Solomon Islands
#> The output will be returned as a list with the elements: 'taxonomyReport', 'SynonymReport', and 'checklistReport'. 
#> These can be accessed using 'output'$taxonomyReport, 'output'$SynonymReport, 'output'$checklistReport, or 'output'$failedReport.

  # Multiple entry example
testQuery <- BeeBDCQuery(
  beeName = c("Homalictus fijiensis", "Homalictus urbanus",
  "Lasioglossum fijiense (Perkins and Cheesman, 1928)"),
  searchChecklist = TRUE,
  printAllSynonyms = TRUE)
#> Starting taxonomy report...
#> Lasioglossum fijiense (Perkins and Cheesman, 1928) is an accpeted name with the taxon id number 32620.
#> Homalictus fijiensis is a synonym of Lasioglossum fijiense (Perkins and Cheesman, 1928) with the taxon id number 32620.
#> Homalictus urbanus is a synonym of Lasioglossum urbanum (Smith, 1879) with the taxon id number 36429.
#>  - 'Lasioglossum fijiense (Perkins and Cheesman, 1928)' has the synonyms: Halictus fijiensis Perkins and Cheesman, 1928, Halictus suvaensis Cockerell, 1929, Homalictus fijiensis (Perkins and Cheesman, 1928), Halictus fijiensis Perkins and Cheesman, 1928, Halictus suvaensis Cockerell, 1929, Homalictus fijiensis (Perkins and Cheesman, 1928)
#>  - 'Homalictus fijiensis' has the synonyms: Halictus fijiensis Perkins and Cheesman, 1928, Halictus suvaensis Cockerell, 1929, Homalictus fijiensis (Perkins and Cheesman, 1928), Halictus fijiensis Perkins and Cheesman, 1928, Halictus suvaensis Cockerell, 1929, Homalictus fijiensis (Perkins and Cheesman, 1928)
#>  - 'Homalictus urbanus' has the synonyms: Halictus urbanus Smith, 1879, Halictus urbanus baudinensis Cockerell, 1905, Halictus cretinicola Friese, 1909, Halictus kesteveni Cockerell, 1912, Halictus hackeriellus Cockerell, 1914, Halictus pavonellus Cockerell, 1915, Halictus olivinus Cockerell, 1922, Halictus urbanus var lomatiae Cockerell, 1922, Halictus microchalceus Cockerell, 1929, Halictus subcarus Cockerell, 1930, Halictus williamsi Cockerell, 1930, Halictus suburbanus Cockerell, 1930, Halictus aponi Cheesman and Perkins, 1939, Halictus aponi var erromangana Cheesman and Perkins, 1939, Homalictus urbanus (Smith, 1879), Homalictus urbanus aponi (Cheesman and Perkins, 1939)
#> Starting checklist report...
#>  - Lasioglossum fijiense (Perkins and Cheesman, 1928) is reportedly found in: 
#> Fiji, Solomon Islands
#>  - Lasioglossum fijiense (Perkins and Cheesman, 1928) is reportedly found in: 
#> Fiji, Solomon Islands
#> The output will be returned as a list with the elements: 'taxonomyReport', 'SynonymReport', and 'checklistReport'. 
#> These can be accessed using 'output'$taxonomyReport, 'output'$SynonymReport, 'output'$checklistReport, or 'output'$failedReport.
  
    # Example way to examine a report from the output list
  testQuery$checklistReport
#> # A tibble: 7 × 23
#>   validName      DiscoverLife_name rNaturalEarth_name shortName DiscoverLife_ISO
#>   <chr>          <chr>             <chr>              <chr>     <chr>           
#> 1 Lasioglossum … Fiji              Fiji               Fiji      FJ              
#> 2 Lasioglossum … Solomon Islands   Solomon Islands    Solomon … BP              
#> 3 Lasioglossum … Australia         Australia          Australia AS              
#> 4 Lasioglossum … Indonesia         Indonesia          Indonesia ID              
#> 5 Lasioglossum … New Caledonia     New Caledonia      New Cale… NC              
#> 6 Lasioglossum … Papua New Guinea  Papua New Guinea   Papua Ne… PP              
#> 7 Lasioglossum … Vanuatu           Vanuatu            Vanuatu   NH              
#> # ℹ 18 more variables: `Alpha-2` <chr>, `Alpha-3` <chr>, official <chr>,
#> #   Source <chr>, matchCertainty <chr>, canonical <chr>,
#> #   canonical_withFlags <chr>, family <chr>, subfamily <chr>, genus <chr>,
#> #   subgenus <lgl>, specificEpithet <chr>, species <chr>, infraspecies <chr>,
#> #   scientificNameAuthorship <chr>, taxon_rank <chr>,
#> #   infraspecificEpithet <chr>, Notes <chr>