Get country names from coordinates

Country names derived from valid geographic coordinates are added to records missing country names. The prefix, jbd_ is meant to distinguish this function from the original bdc::bdc_country_from_coordinates().

jbd_country_from_coordinates(
  data,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  country = "country",
  scale = "large"
)

Arguments

data

A data frame or tibble. This needs to contain, at a minimum, columns with geographical coordinates and country names.

lat

A character string. The column name with latitude in decimal degrees and WGS84. Default = "decimalLatitude".

lon

A character string. The column with longitude in decimal degrees and WGS84. Default = "decimalLongitude".

country

A character string. The column name with the country assignment of each record. Default = "country". If no column name is provided a new column "country" is created.

scale

Passed to rnaturalearth's ne_countries(). Scale of map to return, one of 110, 50, 10 or 'small', 'medium', 'large'. Default = "large".

Value

A tibble of the original input data with country names added for records missing such information and for which a country determination could be made.

Details

This function assigns a country name for records missing such information. Country names are extracted from valid geographic coordinates using a high-quality map of the world (from the rnaturalearth package). No country name is added to records whose coordinates are in the sea.

Examples

if (FALSE) {
x <- data.frame(
  decimalLatitude = c(-22.9834, -39.857030, -17.06811, -46.69778),
  decimalLongitude = c(-69.095, -68.443588, 37.438108, -13.82444),
  country = c("", NA, NA, "Brazil"))

bdc_country_from_coordinates(
  data = x,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  country = "country"
)
}