A country-level checklist of bees from Discover Life

The table contains taxonomic and country information for the bees of the world based on data collated on Discover Life. See beesTaxonomy() for further context.

Usage

data(beesChecklist)

Format

An object of class "tibble"

validName

The valid scientificName as it should occur in the scientificName column.

DiscoverLife_name

The full country name as it occurs on Discover Life.

rNaturalEarth_name

Country name from rnaturalearth's name_long.

shortName

A short version of the country name.

DiscoverLife_ISO

The ISO country name as it occurs on Discover Life.

Alpha-2

Alpha-2 from rnaturalearth.

Alpha-3

Alpha-3 from rnaturalearth.

official

Official country name = "yes" or only a Discover Life name = "no".

Source

A text strign denoting the source or author of the name-country pair.

matchCertainty

Quality of the name's match to the Discover Life checklist.

canonical

The valid species name without scientificNameAuthority.

canonical_withFlags

The validName without the scientificNameAuthority but with Discover Life flags.

family

Bee family.

subfamily

Bee subfamily.

genus

Bee genus.

subgenus

Bee subgenus.

infraspecies

Bee infraSpecificEpithet.

infraspecificEpithet

Bee infraspecificEpithet.

species

Bee specificEpithet.

scientificNameAuthorship

Bee scientificNameAuthorship.

taxon_rank

Rank of the taxon name.

Notes

Discover Life country name notes.

References

This dataset was created using the Discover Life checklist and taxonomy. for further information on download date in file name. Dataset is from the publication: DOREY, J. B., CHESSHIRE, P. R., BOLAÑOS, A. N., O’REILLY, R. L., BOSSERT, S., COLLINS, S. M., LICHTENBERG, E. M., TUCKER, E., SMITH-PARDO, A., FALCON-BRINDIS, A., GUEVARA, D. A., RIBEIRO, B. R., DE PEDRO, D., FISCHER, E., HUNG, J. K.-L., PARYS, K. A., ROGAN, M. S., MINCKLEY, R. L., VELZCO, S. J. E., GRISWOLD, T., ZARRILLO, T. A., SICA, Y., ORR, M. C., GUZMAN, L. M., ASCHER, J., HUGHES, A. C. & COBB, N. S. In review. A globally synthesised and flagged bee occurrence dataset and cleaning workflow. Scientific Data. The checklist data are mostly compiled from Discover Life data, www.discoverlife.org: ASCHER, J. S. & PICKERING, J. 2020. Discover Life bee species guide and world checklist (Hymenoptera: Apoidea: Anthophila). http://www.discoverlife.org/mp/20q?guide=Apoidea_species.

Examples

## Not run:
data(beesChecklist)
head(beesChecklist)
#> # A tibble: 6 × 23
#>   validName      DiscoverLife_name rNaturalEarth_name shortName DiscoverLife_ISO
#>   <chr>          <chr>             <chr>              <chr>     <chr>           
#> 1 Halictus lute… Colombia          Colombia           Colombia  CO              
#> 2 Megachile umb… Kiribati          Kiribati           Kiribati  KR              
#> 3 Megachile umb… Nauru             Nauru              Nauru     NR              
#> 4 Andrena afgha… Afghanistan       Afghanistan        Afghanis… AF              
#> 5 Andrena afzel… Afghanistan       Afghanistan        Afghanis… AF              
#> 6 Andrena albop… Afghanistan       Afghanistan        Afghanis… AF              
#> # ℹ 18 more variables: `Alpha-2` <chr>, `Alpha-3` <chr>, official <chr>,
#> #   Source <chr>, matchCertainty <chr>, canonical <chr>,
#> #   canonical_withFlags <chr>, family <chr>, subfamily <chr>, genus <chr>,
#> #   subgenus <lgl>, specificEpithet <chr>, species <chr>, infraspecies <chr>,
#> #   scientificNameAuthorship <chr>, taxon_rank <chr>,
#> #   infraspecificEpithet <chr>, Notes <chr>
## End(Not run)