A nearly complete taxonomy of bees globally

Contains taxonomic information for the bees of the world. Source of taxonomy is listed under "source" but are mostly derived from the Discover Life website.

data(beesTaxonomy)

Format

An object of class "tibble" ## beesTaxonomy

flags

Flags or comments about the taxon name.

taxonomic_status

Taxonomic status. Values are "accepted" or "synonym"

source

Source of the name.

accid

The id of the accepted taxon name or "0" if taxonomic_status == accepted.

id

The id number for the taxon name.

kingdom

The biological kingdom the taxon belongs to. For bees, kingdom == Animalia.

phylum

The biological phylum the taxon belongs to. For bees, phylum == Arthropoda.

class

The biological class the taxon belongs to. For bees, class == Insecta.

order

The biological order the taxon belongs to. For bees, order == Hymenoptera.

family

The family of bee which the species belongs to.

subfamily

The subfamily of bee which the species belongs to.

tribe

The tribe of bee which the species belongs to.

subtribe

The subtribe of bee which the species belongs to.

validName

The valid scientific name as it should occur in the “scientificName” column in a Darwin Core file.

canonical

The scientificName without the scientificNameAuthority.

canonical_withFlags

The scientificName without the scientificNameAuthority and with Discover Life taxonomy flags.

genus

The genus the bee species belongs to.

subgenus

The subgenus the bee species belongs to.

species

The specific epithet for the bee species.

infraspecies

The infraspecific epithet for the bee addressed.

authorship

The author who described the bee species.

taxon_rank

Rank for the bee taxon addressed in the entry.

valid

Logical for if the name is valid. TRUE == a valid name, FALSE == invalid scientific name.

notes

Additional notes about the name/taxon.

References

This dataset was created using the Discover Life taxonomy. See rawDataWrite.R for further information on download date in file name. Dataset is from the publication: DOREY, J. B., CHESSHIRE, P. R., BOLAÑOS, A. N., O’REILLY, R. L., BOSSERT, S., COLLINS, S. M., LICHTENBERG, E. M., TUCKER, E., SMITH-PARDO, A., FALCON-BRINDIS, A., GUEVARA, D. A., RIBEIRO, B. R., DE PEDRO, D., FISCHER, E., HUNG, J. K.-L., PARYS, K. A., ROGAN, M. S., MINCKLEY, R. L., VELZCO, S. J. E., GRISWOLD, T., ZARRILLO, T. A., SICA, Y., ORR, M. C., GUZMAN, L. M., ASCHER, J., HUGHES, A. C. & COBB, N. S. In review. A globally synthesised and flagged bee occurrence dataset and cleaning workflow. Scientific Data. The taxonomy data are mostly compiled from Discover Life data, www.discoverlife.org: ASCHER, J. S. & PICKERING, J. 2020. Discover Life bee species guide and world checklist (Hymenoptera: Apoidea: Anthophila). http://www.discoverlife.org/mp/20q?guide=Apoidea_species.

Examples

data(beesTaxonomy)
head(beesTaxonomy)
#> # A tibble: 6 × 24
#>   flags taxonomic_status source    accid    id kingdom phylum class order family
#>   <chr> <chr>            <chr>     <dbl> <dbl> <chr>   <chr>  <chr> <chr> <chr> 
#> 1 NA    accepted         Discover…     0     4 Animal… Arthr… Inse… Hyme… Andre…
#> 2 NA    synonym          Discover…     4     5 Animal… Arthr… Inse… Hyme… Andre…
#> 3 NA    accepted         Discover…     0     6 Animal… Arthr… Inse… Hyme… Andre…
#> 4 NA    accepted         Discover…     0     7 Animal… Arthr… Inse… Hyme… Andre…
#> 5 NA    synonym          Discover…     7     8 Animal… Arthr… Inse… Hyme… Andre…
#> 6 NA    accepted         Discover…     0     9 Animal… Arthr… Inse… Hyme… Andre…
#> # ℹ 14 more variables: subfamily <chr>, tribe <chr>, subtribe <chr>,
#> #   validName <chr>, canonical <chr>, canonical_withFlags <chr>, genus <chr>,
#> #   subgenus <chr>, species <chr>, infraspecies <chr>, authorship <chr>,
#> #   taxon_rank <chr>, valid <lgl>, notes <chr>