Build a per-species summary for each and all flags

Takes a flagged dataset and returns the total number of fails (FALSE) per flag (columns starting with ".") and per species. Users may define the column to group the summary by. While it is intended to work with the scientificName column, users may select any grouping column (e.g., country).

flagSummaryTable(
  data = NULL,
  column = "scientificName",
  outPath = OutPath_Report,
  fileName = "flagTable.csv"
)

Arguments

data

A data frame or tibble. The flagged dataset.

column

Character. The name of the column to group by and summarise the failed occurrences. Default = "scientificName".

outPath

A character path. The path to the directory in which the figure will be saved. Default = OutPath_Report.

fileName

Character. The name of the file to be saved, ending in ".csv". Default = "flagTable.csv".

Value

A tibble with a column for each flag column (starting with ".") showing the number of failed (FALSE) occurrences per group. Also shows the (i) total number of records, (ii) total number of failed records, and (iii) the percentage of failed records.

Examples

# Load the toy flagged bee data
data("beesFlagged")

  # Run the function and build the flag table
flagTibble <- flagSummaryTable(data = beesFlagged,
                              column = "scientificName",
                              outPath = paste0(tempdir()),
                              fileName = "flagTable.csv")
#>  - We will flag all columns starting with '.'
#>  - summaryFun:
#> Flagged 81 
#>   The .summary column was added to the database.